requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0
django-cors-headers==3.10.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==3.3
numpy==1.22.3
opencv-python==4.5.5.64
paho-mqtt==1.6.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
